Improve extension runner with plot.
